+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Inspired by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 **Note:** this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2024-03-12
+### Added
+- Add interface for tagging exceptions sent to honeybadger with values from the log context.
+  - `ExceptionHandling.add_honeybadger_tag_from_log_context`
+
+### Changed
+- Require ruby version 2.7 or greater
+
 ## [3.0.1] - 2024-03-07
 ### Changed
 - Allow newer versions of the `psych` gem to be used to avoid conflicts with the default version pulled in by ruby 3.2+